fix(nvim): don't exec CloseFloats command
After updating Neovim to version 0.11.3, I noticed that floating windows stopped closing automatically. This was quite irritating, as the definition window remained open after navigating to a definition or implementation.

Working in split mode also became frustrating, as the floating window did not close in one window when I switched to another. To address this issue, I implemented a CloseFloats function to fix the problem, but I recognized that the root cause was related to the Neovim version.

I then installed the bob Neovim version manager and switched to a previous version, which made my fix unnecessary.